(test): merge passChnge validation into user validation
